add stdint.h

+++ philo/include/philo_data_structures.h
@@ -6,7 +6,7 @@
 /*   By: pmolnar <pmolnar@student.codam.nl>           +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2022/06/07 12:19:42 by pmolnar       #+#    #+#                 */
-/*   Updated: 2022/07/26 13:53:38 by pmolnar       ########   odam.nl         */
+/*   Updated: 2022/07/26 17:38:47 by pmolnar       ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,6 +14,7 @@
 # define PHILO_DATA_STRUCTURES_H
 
 # define MUTEX_SIZE 1
+# include <stdint.h>
 
 typedef struct timeval	t_time;
 typedef pthread_mutex_t	t_mutex;
@@ -27,7 +28,7 @@ typedef struct s_philo
 
 typedef struct s_sim
 {
-	int			attr[5];
+	int16_t		attr[5];
 	t_mutex		mutex[MUTEX_SIZE];
 	t_mutex		*fork;
 	t_philo		*philo;
